pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "cove-pdf-editor"
version = "1.0.0"
description = "Offline PDF editor — edit text, annotate, sign, fill forms, and stamp pages without the cloud"
readme = "README.md"
requires-python = ">=3.10"
license = { text = "MIT" }
authors = [{ name = "Cove" }]
dependencies = [
    "PySide6>=6.6",
    "pypdf>=5.0",
    "pikepdf>=9.0",
    "pypdfium2>=5.0",
    "pdfplumber>=0.11",
    "reportlab>=4.0",
    "Pillow>=10.0",
]

[project.scripts]
cove-pdf-editor = "cove_pdf_editor.__main__:main"

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
cove_pdf_editor = ["assets/*.png", "assets/*.ico"]
